Add files via upload
The HTML and CSS files in this admin dashboard project are used to manage suppliers, users, and the admin profile. The UI has options for updating profiles, vendor and user administration forms, and navigation links. For a unified and user-friendly experience, every area has a similar appearance. Backend integration and dynamic content are possible future improvements.